requirements.txt
altgraph==0.17
future==0.18.2
importlib-metadata==4.5.0
pefile==2021.5.24
pyinstaller==4.3
pyinstaller-hooks-contrib==2021.1
pywin32-ctypes==0.2.0
PyYAML==5.4.1
typing-extensions==3.10.0.0
zipp==3.4.1
